Restyled by yapf

+++ tensorflow/compiler/tests/special_math_test.py
@@ -37,10 +37,8 @@ from tensorflow.python.platform import test
 flags.DEFINE_bool(
     "vary_seed",
     False,
-    (
-        "Whether to vary the PRNG seed unpredictably.  "
-        "With --runs_per_test=N, produces N iid runs."
-    ),
+    ("Whether to vary the PRNG seed unpredictably.  "
+     "With --runs_per_test=N, produces N iid runs."),
 )
 
 NUM_SAMPLES = int(1e3)
@@ -71,7 +69,7 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
                 answer = int(entropy.encode("hex"), 16)
             else:
                 answer = int.from_bytes(entropy, "big")
-            np.random.seed(answer % (2 ** 32 - 1))
+            np.random.seed(answer % (2**32 - 1))
         super(IgammaTest, self).setUp()
 
     # Skip Float64 test on TPU due to missing ops.
@@ -89,13 +87,16 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
             return 2e-2, 1e-7
         return 2e-4, 1e-20
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testLargeXSmallA(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=100.0, high=200.0, size=[NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=0.3, high=1.0, size=[NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=100.0, high=200.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=0.3, high=1.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -104,17 +105,18 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
             actual = sess.run(y)
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testSmallValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
-        ).astype(dtype)
-        a = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
-        ).astype(dtype)
+        x = np.random.uniform(low=np.finfo(dtype).tiny,
+                              high=1.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=np.finfo(dtype).tiny,
+                              high=1.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -122,13 +124,16 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
                 actual = sess.run(_igamma(a, x))
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testMediumValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=1.0, high=100.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=1.0, high=100.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -136,18 +141,17 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
                 actual = sess.run(_igamma(a, x))
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 2e-2, 1e-5), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 2e-2, 1e-5),
+                              (np.float64, 1e-4, 1e-30))
     def testLargeValues(self, dtype, rtol, atol):
         if self.device == "TPU":
             # TODO(b/154908275): Remove this once fixed for large a, x.
             self.skipTest("Skipping test since numerically unstable on TPU.")
         # Test values near zero.
-        x = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
-            dtype
-        )
-        a = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
-            dtype
-        )
+        x = np.random.uniform(low=100.0, high=int(1e4),
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=100.0, high=int(1e4),
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -163,15 +167,11 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
         with self.session():
             with self.test_scope():
                 x = constant_op.constant(
-                    np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(
-                        dtype
-                    )
-                )
+                    np.random.uniform(low=1.0, high=100.0,
+                                      size=[NUM_SAMPLES]).astype(dtype))
                 a = constant_op.constant(
-                    np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(
-                        dtype
-                    )
-                )
+                    np.random.uniform(low=1.0, high=100.0,
+                                      size=[NUM_SAMPLES]).astype(dtype))
 
                 def f(b):
                     return _igamma(b, x)
@@ -187,15 +187,13 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
         with self.session():
             with self.test_scope():
                 x = constant_op.constant(
-                    np.random.uniform(
-                        low=100.0, high=int(1e4), size=[NUM_SAMPLES]
-                    ).astype(dtype)
-                )
+                    np.random.uniform(low=100.0,
+                                      high=int(1e4),
+                                      size=[NUM_SAMPLES]).astype(dtype))
                 a = constant_op.constant(
-                    np.random.uniform(
-                        low=100.0, high=int(1e4), size=[NUM_SAMPLES]
-                    ).astype(dtype)
-                )
+                    np.random.uniform(low=100.0,
+                                      high=int(1e4),
+                                      size=[NUM_SAMPLES]).astype(dtype))
 
                 def f(b):
                     return _igamma(b, x)
@@ -205,7 +203,8 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
                 )
         self.assertLessEqual(max_error, tolerance)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testRandomGammaGradSmallValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
@@ -214,32 +213,31 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
         with self.session() as sess:
             with self.test_scope():
                 x = constant_op.constant(
-                    np.random.uniform(
-                        low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
-                    ).astype(dtype)
-                )
+                    np.random.uniform(low=np.finfo(dtype).tiny,
+                                      high=1.0,
+                                      size=[NUM_SAMPLES]).astype(dtype))
                 a = constant_op.constant(
-                    np.random.uniform(
-                        low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
-                    ).astype(dtype)
-                )
+                    np.random.uniform(low=np.finfo(dtype).tiny,
+                                      high=1.0,
+                                      size=[NUM_SAMPLES]).astype(dtype))
                 gamma_sample_grad = gen_random_ops.random_gamma_grad(a, x)
                 actual_grad = implicit_reparameterization_grad(a, x)
                 gamma_sample_grad, actual_grad = sess.run(
-                    [gamma_sample_grad, actual_grad]
-                )
+                    [gamma_sample_grad, actual_grad])
                 # We do this because the ratio computed in
                 # implicit_reparameterization_grad can very easily result in a NaN due
                 # to the computed numerator and denominator zeroing out.
-                gamma_sample_grad = gamma_sample_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
-                ]
-                actual_grad = actual_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
-                ]
-        self.assertAllClose(actual_grad, gamma_sample_grad, atol=atol, rtol=rtol)
+                gamma_sample_grad = gamma_sample_grad[~np.logical_or(
+                    np.isnan(actual_grad), np.isinf(actual_grad))]
+                actual_grad = actual_grad[~np.logical_or(
+                    np.isnan(actual_grad), np.isinf(actual_grad))]
+        self.assertAllClose(actual_grad,
+                            gamma_sample_grad,
+                            atol=atol,
+                            rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testRandomGammaGradMediumValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
@@ -247,30 +245,26 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
         with self.session() as sess:
             with self.test_scope():
                 x = constant_op.constant(
-                    np.random.uniform(low=1.0, high=10.0, size=[NUM_SAMPLES]).astype(
-                        dtype
-                    )
-                )
+                    np.random.uniform(low=1.0, high=10.0,
+                                      size=[NUM_SAMPLES]).astype(dtype))
                 a = constant_op.constant(
-                    np.random.uniform(low=1.0, high=10.0, size=[NUM_SAMPLES]).astype(
-                        dtype
-                    )
-                )
+                    np.random.uniform(low=1.0, high=10.0,
+                                      size=[NUM_SAMPLES]).astype(dtype))
                 gamma_sample_grad = gen_random_ops.random_gamma_grad(a, x)
                 actual_grad = implicit_reparameterization_grad(a, x)
                 gamma_sample_grad, actual_grad = sess.run(
-                    [gamma_sample_grad, actual_grad]
-                )
+                    [gamma_sample_grad, actual_grad])
                 # We do this because the ratio computed in
                 # implicit_reparameterization_grad can very easily result in a NaN due
                 # to the computed numerator and denominator zeroing out.
-                gamma_sample_grad = gamma_sample_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
-                ]
-                actual_grad = actual_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
-                ]
-        self.assertAllClose(actual_grad, gamma_sample_grad, atol=atol, rtol=rtol)
+                gamma_sample_grad = gamma_sample_grad[~np.logical_or(
+                    np.isnan(actual_grad), np.isinf(actual_grad))]
+                actual_grad = actual_grad[~np.logical_or(
+                    np.isnan(actual_grad), np.isinf(actual_grad))]
+        self.assertAllClose(actual_grad,
+                            gamma_sample_grad,
+                            atol=atol,
+                            rtol=rtol)
 
 
 class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
@@ -281,7 +275,7 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
                 answer = int(entropy.encode("hex"), 16)
             else:
                 answer = int.from_bytes(entropy, "big")
-            np.random.seed(answer % (2 ** 32 - 1))
+            np.random.seed(answer % (2**32 - 1))
         super(IgammacTest, self).setUp()
 
     # Skip Float64 test on TPU due to missing ops.
@@ -300,13 +294,16 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
             return 2e-2, 1e-7
         return 2e-4, 1e-20
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testLargeXSmallA(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=100.0, high=200.0, size=[NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=0.3, high=1.0, size=[NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=100.0, high=200.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=0.3, high=1.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
@@ -315,17 +312,18 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
             actual = sess.run(y)
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testSmallValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
-        ).astype(dtype)
-        a = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
-        ).astype(dtype)
+        x = np.random.uniform(low=np.finfo(dtype).tiny,
+                              high=1.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=np.finfo(dtype).tiny,
+                              high=1.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
@@ -333,13 +331,16 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
                 actual = sess.run(_igammac(a, x))
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11),
+                              (np.float64, 1e-4, 1e-30))
     def testMediumValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=1.0, high=100.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=1.0, high=100.0,
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
@@ -347,17 +348,16 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
                 actual = sess.run(_igammac(a, x))
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 2e-2, 1e-5), (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 2e-2, 1e-5),
+                              (np.float64, 1e-4, 1e-30))
     def testLargeValues(self, dtype, rtol, atol):
         if self.device == "TPU":
             self.skipTest("Skipping test since numerically unstable on TPU.")
         # Test values near zero.
-        x = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
-            dtype
-        )
-        a = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
-            dtype
-        )
+        x = np.random.uniform(low=100.0, high=int(1e4),
+                              size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=100.0, high=int(1e4),
+                              size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
